backend/requirements.txt
Flask==1.1.2
Flask-Cors
Flask-JWT-Simple==0.0.3
Flask-RESTful==0.3.8
logger==1.4
PyJWT==1.7.1
python-crfsuite==0.9.7
python-engineio==3.13.2
sklearn_crfsuite==0.3.6
pyspellchecker
tornado==5.1.1
Werkzeug==1.0.1
